refactor(virrun): file-organization sweep over the cli, configuration, source and virrun services
Every inline options shape the CLI formatters took is a model (the
cache listing, the doctor report, the banner and result lines, the run
options), and the two location-shaped ones are `Pick`s of the snapshot
and prepare locations they already mirror. The private helpers become
services of their own: the keyed-tier line, the SGR map, the two cache
clean writers, the tier reader, the doctor's row and summary
formatters with their status maps, the five doctor probes with their
two probe readers and the bubblewrap floor, the version-triple parser,
the package-scripts reader and the backend factory map.
