+++ packages/virrun/src/models/cli/DoctorReport.ts
@@ -1,0 +1,6 @@
+import type { DiagnosticCheck } from "#src/models/cli/DiagnosticCheck";
+
+export interface DoctorReport {
+  checks: readonly DiagnosticCheck[];
+  platform: string;
+}
