+++ packages/virrun/src/models/cli/RunVirrunCommandOptions.ts
@@ -1,0 +1,5 @@
+import type { ExecutionMode } from "#src/models/virrun/ExecutionMode";
+
+export interface RunVirrunCommandOptions {
+  mode: ExecutionMode;
+}
