+++ packages/virrun/src/models/cli/VirrunBanner.ts
@@ -1,0 +1,7 @@
+import type { BackendType } from "#src/models/virrun/BackendType";
+
+export interface VirrunBanner {
+  backend: BackendType;
+  command: readonly string[];
+  nodeVersion: string;
+}
