+++ packages/virrun/src/models/cli/VirrunResult.ts
@@ -1,0 +1,5 @@
+export interface VirrunResult {
+  command: readonly string[];
+  durationMs: number;
+  exitCode: number;
+}
