+++ packages/virrun/src/services/cli/doctor/DiagnosticStatusLabelMap.ts
@@ -1,0 +1,7 @@
+import { DiagnosticStatus } from "#src/models/cli/DiagnosticStatus";
+// The status column word for each outcome; padded to the widest ("MISSING") so the note column aligns.
+export const DiagnosticStatusLabelMap: Record<DiagnosticStatus, string> = {
+  [DiagnosticStatus.Missing]: "MISSING",
+  [DiagnosticStatus.NotApplicable]: "n/a",
+  [DiagnosticStatus.Ok]: "ok",
+};
