+++ packages/virrun/src/services/cli/doctor/constants.ts
@@ -1,0 +1,2 @@
+// The oldest bubblewrap exposing `--overlay-src` / `--tmp-overlay` (the RAM-overlay flags the os backend needs).
+export const MINIMUM_BUBBLEWRAP_VERSION = "0.10.0";
