+++ packages/virrun/src/services/cli/run/checkIsVersionAtLeast.ts
@@ -1,13 +1,4 @@
-// Extracts the leading major.minor.patch triple from a version string — accepts a bare "0.11.1" or a
-// `bwrap --version` line ("bubblewrap 0.11.1"); undefined when no triple is present so a garbled `--version` reads
-// As unknown, never as satisfying the gate.
-const VERSION_REGEX = /(?<major>\d+)\.(?<minor>\d+)\.(?<patch>\d+)/u;
-
-const parseVersionTriple = (version: string): [number, number, number] | undefined => {
-  const groups = VERSION_REGEX.exec(version)?.groups;
-  if (groups === undefined) return undefined;
-  else return [Number(groups.major), Number(groups.minor), Number(groups.patch)];
-};
+import { parseVersionTriple } from "#src/services/cli/run/parseVersionTriple";
 // Whether `version` is >= `minimum`, compared field-by-field on the major.minor.patch triple. An unparseable
 // Version (or minimum) reads as below the minimum — unknown is treated as unsupported.
 export const checkIsVersionAtLeast = (version: string, minimum: string): boolean => {
