+++ packages/virrun/src/services/cli/run/getCommandNotFoundHint.ts
@@ -1,25 +1,14 @@
 import { Color } from "#src/models/cli/Color";
 import { colorize } from "#src/services/cli/color/colorize";
 import { formatVirrunLine } from "#src/services/cli/format/formatVirrunLine";
-import { parseMachineJson } from "#src/services/exec/util/parseMachineJson";
-import { getResult, takeOne } from "@esposter/shared";
-import { readFileSync } from "node:fs";
-import { join } from "node:path";
-import { z } from "zod";
+import { readPackageScripts } from "#src/services/cli/run/readPackageScripts";
+import { takeOne } from "@esposter/shared";
 // The sandbox prints `bwrap: execvp <cmd>: No such file or directory` and node's ENOENT surfaces as `spawn <cmd> ENOENT`
 // When the intended executable is missing from PATH — both are the same user error: a package script (or a typo)
 // Was passed where virrun expects a real executable, so the sandbox-setup message misleads. When the missing token
 // Is a package.json script in `cwd`, point at the `virrun -- pnpm <script>` form that actually resolves.
 const COMMAND_NOT_FOUND_REGEX =
   /(?:execvp (?<execvpCommand>\S+): No such file or directory|spawn (?<spawnCommand>\S+) ENOENT)/u;
-const packageScriptsSchema = z.object({ scripts: z.record(z.string(), z.string()).default({}) });
-
-const readPackageScripts = (cwd: string): string[] =>
-  getResult(() => packageScriptsSchema.parse(parseMachineJson(readFileSync(join(cwd, "package.json"), "utf8")))).match(
-    (packageJson) => Object.keys(packageJson.scripts),
-    () => [],
-  );
-
 export const getCommandNotFoundHint = (
   command: readonly string[],
   errorMessage: string,
